Displays cargo color on ShuffleBoard
When shuffleboard is open, the Boolean block called "Cargo Color" will appear. You need to set the colors to red when true and blue when false in "edit properties". You get to this by right clicking on the Boolean block.  When you close Shuffleboard, save the layout and it will save the true and false colors.

Cargo Color also will default to blue (false) until a different color is picked up. This can be changed in Robot.Java on line 37